update flow order

+++ src/main/java/box/bookstorebe/document/payment/PaymentDocument.java
@@ -9,6 +9,8 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.mongodb.core.mapping.Field;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -23,7 +25,7 @@ public class PaymentDocument {
     private String paymentTime;
 
     @Field(name = "total_price")
-    private int totalPrice;
+    private BigDecimal totalPrice;
 
     @Field(name = "transactionId")
     private String transactionId;
